requirements.txt
streamlit==1.33.0
fastapi==0.110.1
starlette==0.37.2
pydantic==2.7.0
pydantic_core==2.18.1
itsdangerous==2.1.2
python-dotenv==1.0.1
uvicorn==0.29.0

